Complete the unit and staging test cases for the Fund Me contract

+++ contracts/PriceConverter.sol
@@ -4,21 +4,36 @@ pragma solidity ^0.8.8;
 import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 library PriceConverter {
-    function getDecimals(AggregatorV3Interface priceFeed) internal view returns (uint256) {
+    function getDecimals(AggregatorV3Interface priceFeed)
+        internal
+        view
+        returns (uint256)
+    {
         return priceFeed.decimals();
     }
 
-    function getLatestPrice(AggregatorV3Interface priceFeed) internal view returns (uint256) {
+    function getLatestPrice(AggregatorV3Interface priceFeed)
+        internal
+        view
+        returns (uint256)
+    {
         uint256 decimalsLeft = uint256(18) - getDecimals(priceFeed);
-        (,int price,,,) = priceFeed.latestRoundData();
-        return uint256(price) * 10 ** decimalsLeft;
+        (, int256 price, , , ) = priceFeed.latestRoundData();
+        return uint256(price) * 10**decimalsLeft;
     }
 
-    function getVersion(AggregatorV3Interface priceFeed) internal view returns (uint256) {
+    function getVersion(AggregatorV3Interface priceFeed)
+        internal
+        view
+        returns (uint256)
+    {
         return priceFeed.version();
     }
 
-    function getConversionRate(AggregatorV3Interface priceFeed, uint256 ethAmount) internal view returns (uint256) {
+    function getConversionRate(
+        AggregatorV3Interface priceFeed,
+        uint256 ethAmount
+    ) internal view returns (uint256) {
         uint256 ethPriceInUSD = getLatestPrice(priceFeed);
         uint256 ethAmountInUSD = (ethAmount * ethPriceInUSD) / 1e18;
         return ethAmountInUSD;
